Bank Application: - Account 3 Types of accounts available - Checking, Savings & MaxiSavings Interest is compounded daily
- Customer
New customer is set up using a name and customer ID
Customer can have any number of accounts
Total Interest earned is sum of compounded daily interest on each account

- Bank
Has an array of customers
Total interest paid is sum of interest earned by all the customers

Addition Features
- Customers can transfer between accounts

- Interest rate is compounded daily
This is done by grouping transactions by date.
For each day:
Interest = (BalanceOnDay(1 + InterestRate/365)) - BalanceOnDay
This interest is then added to the calculation of interest on the next day.

If no transactions happen for x number of days
Interest = (BalanceOnDay(1 + InterestRate/365)^x) - BalanceOnDay

- MaxiSavings interest rate of 0.01 if withdrawal within the last 10 days else 5%
If a withdrawal happens then for the next 10 days interest rate is set to 0.001
Example

Deposit on 3rd feb
withdrawal on 5th feb
deposit on 18th feb

Interest would be compounded daily interest between:
3rd - 5th  - 5% Interest
5th - 15th Feb 0.01 % Interest
15th - 18th - 5%

This is because interest is compounded daily so for days between 5 -15th there has been a withdrawal with the past 10 days.
Total interest earned is the sum of the interest 3rd - 18th feb.